refactor(bootykey): move BootyKeySupplier to types package and update imports
The BootyKeySupplier class was moved from behaviours/profilechanger to types/suppliers package to better organize the code structure. All related imports were updated accordingly. Also bumped plugin version to 2.8.0 beta 1.
